clean：ファイルの切り分け、キャメルケースの修正
