+++ src/features/expense/history/const/mockData.ts
@@ -55,61 +55,67 @@ export const expenseData: ExpenseType[] = [
   {
     id: 'h7i7k719-0jdh-1lei-gmfh-8i9j0k1l2m3n',
     userId: 'user002',
-    requestDate: new Date('2025-05-05T00:00:00.000Z'), 
+    requestDate: new Date('2025-05-05T00:00:00.000Z'),
     expenseType: 'General',
     amount: 1800,
     description: 'チームのランチミーティング',
     statusId: 'Pending',
     receiptUrl: 'https://example.com/receipts/008.pdf',
-  },{
+  },
+  {
     id: 'h8i7k719-0jdh-1lei-gmfh-8i9j0k1l2m3n',
     userId: 'user002',
-    requestDate: new Date('2025-05-05T00:00:00.000Z'), 
+    requestDate: new Date('2025-05-05T00:00:00.000Z'),
     expenseType: 'General',
     amount: 1800,
     description: 'チームのランチミーティング',
     statusId: 'Pending',
     receiptUrl: 'https://example.com/receipts/008.pdf',
-  },{
+  },
+  {
     id: 'h9i7k719-0jdh-1lei-gmfh-8i9j0k1l2m3n',
     userId: 'user002',
-    requestDate: new Date('2025-05-05T00:00:00.000Z'), 
+    requestDate: new Date('2025-05-05T00:00:00.000Z'),
     expenseType: 'General',
     amount: 1800,
     description: 'チームのランチミーティング',
     statusId: 'Pending',
     receiptUrl: 'https://example.com/receipts/008.pdf',
-  },{
+  },
+  {
     id: 'g1i7k719-0jdh-1lei-gmfh-8i9j0k1l2m3n',
     userId: 'user002',
-    requestDate: new Date('2025-05-05T00:00:00.000Z'), 
+    requestDate: new Date('2025-05-05T00:00:00.000Z'),
     expenseType: 'General',
     amount: 1800,
     description: 'チームのランチミーティング',
     statusId: 'Pending',
     receiptUrl: 'https://example.com/receipts/008.pdf',
-  },{
+  },
+  {
     id: 'g2i7k719-0jdh-1lei-gmfh-8i9j0k1l2m3n',
     userId: 'user002',
-    requestDate: new Date('2025-05-05T00:00:00.000Z'), 
+    requestDate: new Date('2025-05-05T00:00:00.000Z'),
     expenseType: 'General',
     amount: 1800,
     description: 'チームのランチミーティング',
     statusId: 'Pending',
     receiptUrl: 'https://example.com/receipts/008.pdf',
-  },{
+  },
+  {
     id: 'g3i7k719-0jdh-1lei-gmfh-8i9j0k1l2m3n',
     userId: 'user002',
-    requestDate: new Date('2025-05-05T00:00:00.000Z'), 
+    requestDate: new Date('2025-05-05T00:00:00.000Z'),
     expenseType: 'General',
     amount: 1800,
     description: 'チームのランチミーティング',
     statusId: 'Pending',
     receiptUrl: 'https://example.com/receipts/008.pdf',
-  },{
+  },
+  {
     id: 'g4i7k719-0jdh-1lei-gmfh-8i9j0k1l2m3n',
     userId: 'user002',
-    requestDate: new Date('2025-05-05T00:00:00.000Z'), 
+    requestDate: new Date('2025-05-05T00:00:00.000Z'),
     expenseType: 'General',
     amount: 1800,
     description: 'チームのランチミーティング',
@@ -122,28 +128,34 @@ export const routeInfoData: RouteInfoType[] = [
   {
     id: 'r1a2b3c4-5d6e-7f8g-9h0i-1j2k3l4m5n6o',
     expenseRequestId: 'a3b8f042-3c6e-4e7b-9f8a-1b2c3d4e5f6g',
-    routeDetails: {
-      from: '東京',
-      to: '大阪',
-      fare: 12000,
-    },
+    routeDetails: [
+      {
+        from: '東京',
+        to: '大阪',
+        fare: 12000,
+      },
+    ],
   },
   {
     id: 'r2b3c4d5-6e7f-8g9h-0i1j-2k3l4m5n6o7p',
     expenseRequestId: 'f4g5i597-8hbf-9jcg-ekdf-6g7h8i9j0k1l',
-    routeDetails: {
-      from: '品川',
-      to: '大手町',
-      fare: 3000,
-    },
+    routeDetails: [
+      {
+        from: '品川',
+        to: '大手町',
+        fare: 3000,
+      },
+    ],
   },
   {
     id: 'r3c4d5e6-7f8g-9h0i-1j2k-3l4m5n6o7p8q',
     expenseRequestId: 'i7j8l820-1kei-2mfj-hngi-9j0k1l2m3n4o',
-    routeDetails: {
-      from: '東京',
-      to: '名古屋',
-      fare: 8000,
-    },
+    routeDetails: [
+      {
+        from: '東京',
+        to: '名古屋',
+        fare: 8000,
+      },
+    ],
   },
 ];
